feature: ui for no data
